refactor: make collectCellsVals be module, use the last state of table for sorting
